Update popup.css
Changed the textarea so that it does not extend outside the editor div. You can now see the "gripper" in the lower right corner to resize the window. Changed the foldertree items spacing, added a border to foldertree, and some spacing between the editor div. Rounded the corners on the context menu, added a shadow and made the text a bit smaller.